moving STT RecognizeStream to it's own file
